refactor: cleaning child feature and creating test account before all tests
