+++ test_e2e/src/features/child.feature
@@ -1,57 +1,33 @@
 Feature: Crianca
 
-  Scenario Outline: Adiciona presente
-    Given acesso a pagina de login do sistema
-    When acesso o sistema como usuario admin
-    And pesquiso por "<filter_query>" na tabela de criancas
-    And visualizo o detalhe da crianca
-    And seleciono todos items da crianca
-    And preencho o formulario com "<godparent_name>" "<godparent_phone>"
-    And clico no botao de adicionar
-    And clico no botao de salvar
+  @child
+  Scenario Outline: Apadrinhando todos os itens da crianca
+    Given acesso o sistema com usuario "<user_type>"
+    When apadrinho todos os items de uma crianca
     Then as informacoes devem aparecer na tabela
     And uma janela de aviso de sucesso deve aparecer
-    And devo excluir o registro criado
     Examples:
-      | filter_query | godparent_name    | godparent_phone |
-      | nogueira     | Vinicius Avansini | (19) 98119-2732 |
-      # TODO: adicionar mais algum exemplo ou remover
+      | user_type |
+      | admin     |
+      | regular   |
 
-
-  Scenario Outline: Altera presente
-    Given acesso a pagina de login do sistema
-    When acesso o sistema como usuario admin
-    And pesquiso por "<filter_query>" na tabela de criancas
-    And visualizo o detalhe da crianca
-    And seleciono todos items da crianca
-    And preencho o formulario com "<godparent_name>" "<godparent_phone>"
-    And clico no botao de adicionar
-    And clico no botao de salvar
-    And clico na linha do padrinho
-    And preencho o formulario com "Vini2" "(xx) XXXXX-XXXX"
-    And clico no botao de atualizar
-    And clico no botao de salvar
+  @child
+  Scenario Outline: Altera items selecionados do padrinho
+    Given que uma crianca eh apadrinhada com usuario "<user_type>"
+    When atualizo as informacoes do padrinho
     Then uma janela de aviso de sucesso deve aparecer
-    And devo excluir o registro criado
     Examples:
-      | filter_query | godparent_name    | godparent_phone |
-      | nogueira     | Vinicius Avansini | (19) 98119-2732 |
+      | user_type |
+      | admin     |
+      | regular   |
 
-
-  Scenario Outline: Exclui presente
-    Given acesso a pagina de login do sistema
-    When acesso o sistema como usuario admin
-    And pesquiso por "<filter_query>" na tabela de criancas
-    And visualizo o detalhe da crianca
-    And seleciono todos items da crianca
-    And preencho o formulario com "<godparent_name>" "<godparent_phone>"
-    And clico no botao de adicionar
-    And clico no botao de salvar
-    And clico no botao excluir padrinho
-    And clico no botao de salvar
+  @child
+  Scenario Outline: Exclui padrinho de uma crianca
+    Given que uma crianca eh apadrinhada com usuario "<user_type>"
+    When excluo o padrinho da crianca
     Then uma janela de aviso de sucesso deve aparecer
     And nao deve haver informacoes na tabela de items apadrinhados
-    And devo excluir o registro criado
     Examples:
-      | filter_query | godparent_name    | godparent_phone |
-      | nogueira     | Vinicius Avansini | (19) 98119-2732 |
+      | user_type |
+      | admin     |
+      | regular   |
