+++ test_e2e/src/repository/queries.py
@@ -1,3 +1,4 @@
+import uuid
 import psycopg2
 from psycopg2 import Error
 
@@ -15,19 +16,10 @@ def open_connection():
 
 def get_child_by_name(child_name):
     try:
-        # Connect to an existing database
         connection = open_connection()
-
-        # Create a cursor to perform database operations
         cursor = connection.cursor()
-        # Print PostgreSQL details
-        print("PostgreSQL server information")
-        print(connection.get_dsn_parameters(), "\n")
-        # Executing a SQL query
         cursor.execute("SELECT * FROM children WHERE name = %s;", (child_name,))
-        # Fetch result
         record = cursor.fetchone()
-        print("You are connected to - ", record, "\n")
         return record
 
     except (Exception, Error) as error:
@@ -36,67 +28,63 @@ def get_child_by_name(child_name):
         if connection:
             cursor.close()
             connection.close()
-            print("PostgreSQL connection is closed")
 
 
 def delete_godparent(child_id):
     try:
-        # Connect to an existing database
         connection = open_connection()
-
-        # Create a cursor to perform database operations
         cursor = connection.cursor()
         cursor.execute('DELETE FROM god_parents WHERE god_parents."ChildId" = %s;', (child_id,))
         connection.commit()
-        count = cursor.rowcount
-        print(count, "Record deleted successfully")
-
     except (Exception, Error) as error:
         print("Error while connecting to PostgreSQL", error)
     finally:
         if connection:
             cursor.close()
             connection.close()
-            print("PostgreSQL connection is closed")
 
 
 def delete_user(user_login):
     try:
-        # Connect to an existing database
         connection = open_connection()
-
-        # Create a cursor to perform database operations
         cursor = connection.cursor()
         cursor.execute("DELETE FROM users WHERE login = %s;", (user_login,))
         connection.commit()
-        count = cursor.rowcount
-        print(count, "Record deleted successfully ")
-
     except (Exception, Error) as error:
         print("Error while connecting to PostgreSQL", error)
     finally:
         if connection:
             cursor.close()
             connection.close()
-            print("PostgreSQL connection is closed")
 
 
 def reset_user_password(user_login):
     try:
-        # Connect to an existing database
         connection = open_connection()
-
-        # Create a cursor to perform database operations
         cursor = connection.cursor()
         cursor.execute("UPDATE users SET password = %s WHERE login = %s;", ('123', user_login,))
         connection.commit()
-        count = cursor.rowcount
-        print(count, "Record updated successfully ")
-
     except (Exception, Error) as error:
         print("Error while connecting to PostgreSQL", error)
     finally:
         if connection:
             cursor.close()
             connection.close()
-            print("PostgreSQL connection is closed")
+
+
+def insert_user(login, name, password, user_type_id):
+    try:
+        connection = open_connection()
+        cursor = connection.cursor()
+        cursor.execute(
+            "INSERT INTO users(id, login, name, password, user_type_id, is_deleted)"
+            "VALUES (%s, %s, %s, %s, %s, %s)",
+            (str(uuid.uuid4()), login, name, password, user_type_id, False)
+        )
+        connection.commit()
+    except (Exception, Error) as error:
+        print("Error while connecting to PostgreSQL", error)
+    finally:
+        if connection:
+            cursor.close()
+            connection.close()
